requirements.txt
build==1.0.3
nbformat==5.9.2
openai==1.6.1
python-dotenv==1.0.0